refactor(channels): split ingress error factory out of ingress-monitor for max-lines
Move createChannelIngressError into ingress-errors.ts so ingress-monitor.ts
stays under the 700 counted-line lint budget after the terminal-outcome fix.
The message barrel and plugin-sdk keep re-exporting the same symbol, so
channel plugins see no import change.
